Combined Chaos Program
In Chapter 1 we had a program for making a chaotic program at the end
it provided different ideas about things to be done to the chaos
program.  Instead of individual programs for each possible, I put them
all into a single program, using functions to logically separate the
the various needs.

Upgrade ideas: move the formulas as individual functions instead of
having a lot of repeats, Allow the user to be able to try other options
without needing to reload, add in ways to handle erroneous user inputs