Project 2A
Creates 5 graphic shapes and allows the user to click 5 times to
generate 5 different fill and outline colors and a final click to close
the window.  Ideas for upgrading would be to loop number of times asked
for by the user, allow them to save the shapes and their colors codes.